Fix unittests && Provide `login` in SampleFrontendTestBase already
- Stop randomly switching `has_to_pay`, which broke the unittests

- Use a logged in user for `usertraffic` since flasks test_client has no ip

- Provide `login` and `logout` functionality in `SampleFrontendTestBase`
  and not only `SampleAuthenticatedTestBase`

- Use the current directory instead of hardcoding `/home/sipa/sipa`
